make tests pass
The main problem with the tests not passing was the fact
that when you created a new SchoolDay object, it reused the attributes
of the previous one.

That was not the only issue, some minor issues too and some error in
schedule.json.

This should make it almost ready for merge into master