improve list reasignment bug workaround
I just read a Reddit comment with bad practices, and one of those was
the annoying bug I was facing, that the lessons argument would start out
at its previous value when creating a new object. Now I kinda know why.

So I linked it in a comment, and improved the workaround a bit as
suggested by the comment.

However the part in the else statement is maybe needed, maybe not. The
comment didn't mention that as it said about only mutating, not copying.

So the else part still has the previous workaround.